IFA scenario (#75)
    Category: implementation
    JIRA issue: MIC-4747
    Research reference: scenarios

Changes and notes
Replace universal BEP with IFA scenario. Only change to intervention.py is checking if we are in IFA scenario, in which case we assign everyone to the IFA intervention. Otherwise, go through same logic as before.

Verification and Testing
Checked that we have everyone on IFA in IFA scenario and that this is the same way that mms behaves.